GEN-4 Fixed Next layout & image processing
