+++ src/components/Image/index.tsx
@@ -1,0 +1,2 @@
+import Image from './Image';
+export default Image;
