+++ src/components/Layout/Layout.module.scss
@@ -23,7 +23,3 @@
   color: blue;
 
 }
-
-.logo {
-  cursor: pointer;
-}
