+++ src/pages/_app.tsx
@@ -1,0 +1,19 @@
+import type { NextPage } from 'next'
+import type { AppProps } from 'next/app'
+import type { ReactElement, ReactNode } from 'react'
+
+import '../styles/globals.scss';
+import Layout from '@/components/Layout';
+
+type NextPageWithLayout = NextPage & {
+  getLayout?: (page: ReactElement) => ReactNode
+}
+
+type AppPropsWithLayout = AppProps & {
+  Component: NextPageWithLayout
+}
+
+export default function App({ Component, pageProps }: AppPropsWithLayout) {
+  const getLayout = Component.getLayout || Layout.getLayout;
+  return getLayout(<Component {...pageProps} />)
+}
