+++ src/pages/wishlist.tsx
@@ -1,0 +1,18 @@
+import type { NextPage } from 'next';
+
+const Whishlist: NextPage = () => {
+  return (
+    <>
+      <h1>Список желаний</h1>
+      <article>
+        <p>
+          Lorem ipsum dolor sit amet consectetur adipisicing elit. Aliquid eveniet quis est? Quod ullam quos soluta repellendus nihil dolor, mollitia reiciendis aperiam dolores tempora eveniet numquam exercitationem ad at porro!
+          Lorem ipsum dolor sit amet consectetur adipisicing elit. Animi eveniet rem quas repudiandae fuga voluptates iure dicta, facilis in velit aliquid nesciunt pariatur, officiis temporibus consequatur adipisci. Omnis, aut impedit?
+          Lorem, ipsum dolor sit amet consectetur adipisicing elit. Deserunt vitae maxime harum repudiandae, quo nostrum odit. Itaque, reprehenderit iste blanditiis impedit veritatis commodi voluptatem accusamus ad reiciendis quae, quam est?
+        </p>
+      </article>
+    </>
+  )
+}
+
+export default Whishlist;
